--- a/engine/src/main/java/org/hibernate/validator/internal/engine/valueextraction/ValueExtractorManager.java
+++ b/engine/src/main/java/org/hibernate/validator/internal/engine/valueextraction/ValueExtractorManager.java
@@ -191,6 +191,10 @@ public class ValueExtractorManager {
 		}
 	}
 
+	public void clear() {
+		valueExtractorResolver.clear();
+	}
+
 	/**
 	 * Runs the given privileged action, using a privileged block if required.
 	 * <p>