* https://github.com/hibernate/hibernate-validator/pull/878
 * https://hibernate.atlassian.net/browse/HV-1444