@marko-bekhta 2017-11-27
and some numbers:
```
with changes applied:
# Run complete. Total time: 00:01:51

Benchmark                                                                                   Mode  Cnt    Score    Error   Units
CascadedValidation.testCascadedValidation                                                  thrpt   20  241.063 ±  8.683  ops/ms
CascadedWithLotsOfItemsAndMoreConstraintsValidation.testCascadedValidationWithLotsOfItems  thrpt   20  613.279 ± 10.496   ops/s
CascadedWithLotsOfItemsValidation.testCascadedValidationWithLotsOfItems                    thrpt   20  821.396 ±  9.824   ops/s


  Raw validation time:        1096ms {12342}
  Raw parsing time:              0ms {0}
  First parse and validate:      0ms {0}
  Multithreaded validation:      0ms 
[INFO] Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 3.192 s - in org.apache.bval.bench.BenchTest

======================================================
current (without last commit on potentially CMD):
# Run complete. Total time: 00:01:52

Benchmark                                                                                   Mode  Cnt    Score    Error   Units
CascadedValidation.testCascadedValidation                                                  thrpt   20  212.603 ± 26.591  ops/ms
CascadedWithLotsOfItemsAndMoreConstraintsValidation.testCascadedValidationWithLotsOfItems  thrpt   20  592.819 ± 20.631   ops/s
CascadedWithLotsOfItemsValidation.testCascadedValidationWithLotsOfItems                    thrpt   20  785.133 ± 13.543   ops/s


  Raw validation time:         884ms {12342}
  Raw parsing time:              0ms {0}
  First parse and validate:      0ms {0}
  Multithreaded validation:      0ms 
[INFO] Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 3.143 s - in org.apache.bval.bench.BenchT

======================================================
master:

# Run complete. Total time: 00:01:53

Benchmark                                                                                   Mode  Cnt    Score    Error   Units
CascadedValidation.testCascadedValidation                                                  thrpt   20  226.021 ± 37.921  ops/ms
CascadedWithLotsOfItemsAndMoreConstraintsValidation.testCascadedValidationWithLotsOfItems  thrpt   20  607.903 ±  8.669   ops/s
CascadedWithLotsOfItemsValidation.testCascadedValidationWithLotsOfItems                    thrpt   20  860.026 ± 11.842   ops/s



  Raw validation time:         903ms {12342}
  Raw parsing time:              0ms {0}
  First parse and validate:      0ms {0}
  Multithreaded validation:      0ms 
[INFO] Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 3.113 s - in org.apache.bval.bench.BenchTest
```