@gsmet 2017-11-27
Mmmh, did you mess up with the BV benchmark number or is your last version slower than everything and the current version faster than master?

it looks weird to me.