@marko-bekhta 2017-11-27
it's possible that I did ... So I've run the tests again :)
```
current master:
  Raw validation time:         946ms {12342}
  Raw parsing time:              0ms {0}
  First parse and validate:      0ms {0}
  Multithreaded validation:      0ms 
[INFO] Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 3.263 s - in org.apache.bval.bench.BenchTest

with all changes:
  Raw validation time:        1056ms {12342}
  Raw parsing time:              0ms {0}
  First parse and validate:      0ms {0}
  Multithreaded validation:      0ms 
[INFO] Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 3.126 s - in org.apache.bval.bench.BenchTest
```
they are pretty close, and I think that it's faster with these latest changes because in case we have `NonContainerCascadingMetaData` we don't call `NonContainerCascadingMetaData#addRuntimeLegacyCollectionSupport` anymore, which is still called in the code form master (well there is a similar call but in case of `NonContainerCascadingMetaData` it will just return this).