@marko-bekhta 2017-11-27
Also about that BV benchmark.... I think it's possible to convert it to JMH test. We would need to have two modules though. One to generate beans and in the other one we can write a JMH benchmark using those beans.