@gsmet 2017-11-27
@marko-bekhta it's slower in your "with all changes" case. Note that the unit here is ms. So the lower the better.

I don't have anything against moving the existing code to a submodule and adding a new module with some JMH testing. It would be nice if we had the same possibilities than the existing benchmark: the ability to test the raw parsing speed and the raw validation speed (having both mixed is of no use IMHO).