@marko-bekhta 2017-11-27
Hah, yes ..... It's in ms !!! :smiley: but I did try to build a project with `git` today instead of `mvn` so I'm not surprised that I missed the units of a benchmark...
And I just took the "best" numbers from a couple of runs. 

Here I've run it once more and copied 5 results for each:
```
master:
Raw validation time:         900ms {12342}
Raw validation time:         791ms {12342}
Raw validation time:         825ms {12342}
Raw validation time:         886ms {12342}
Raw validation time:         901ms {12342}

changes:
Raw validation time:         834ms {12342}
Raw validation time:         834ms {12342}
Raw validation time:         866ms {12342}
Raw validation time:        1005ms {12342}
Raw validation time:        1123ms {12342}
```