@gsmet 2017-11-27
Mmmmf. I think the only conclusion we can draw is that we really need to move to JMH :).