@marko-bekhta 2017-11-28
Finally had a chance to spent some time on this. I've edited the last commit and force-pushed it. 
I tried to add explanations in the comments with some examples. Hope those would make it easier to read and understand the changes.