@marko-bekhta 2017-11-30
It's not related to VE directly, just encountered one more case where annotations are not read correctly, here's an example:
```java
@Test
public void testTypeVariableAnnotationsIssue() throws Exception {
	class Barrr {
	}
	class MapTest {
		private final Map<@Valid List<@NotBlank String>, @NotNull @Valid Barrr> container;

		MapTest(Map<List<String>, Barrr> container) {
			this.container = container;
		}
	}

	AnnotatedType[] annotatedActualTypeArguments = ( (AnnotatedParameterizedType) MapTest.class.getDeclaredFields()[0].getAnnotatedType() ).getAnnotatedActualTypeArguments();
	assertEquals( annotatedActualTypeArguments.length, 2 );
	assertEquals( annotatedActualTypeArguments[0].getAnnotations().length, 1 );
	assertEquals( annotatedActualTypeArguments[1].getAnnotations().length, 2 ); // this part fails
}
```
Last assert fails if the type is a local one (`Barrr`). If it is replaced with for example `Integer` - everything is ok.