@gsmet 2017-12-04
@marko-bekhta might be nice to add an issue with a test case to our JDK issues milestone.