@gsmet 2017-12-06
Here is the situation:
 * I want to make a last pass to change a few method names and get the overall thing easier to understand
 * Marko is working in parallel on the conversion of the BV benchmark to JMH so that we can have stable results and determine the real consequences of this patch

I'll ping you if I think an additional review is required but we should really wait for the final changes and the benchmark results.