@gsmet 2018-02-01
@marko-bekhta I'm finally done with this PR. We can check that the performances are not too bad, now :).

Will you do it or should I?